[M2] Change event modal logic to use routes and query
